Include XADMaster in Xee, update ArchiveSource. Still has some issues with linking.

+++ XeeDirectorySource.h
@@ -81,6 +81,6 @@
 -(BOOL)matchesObject:(id)obj;
 
 -(BOOL)isEqual:(XeeDirectoryEntry *)other;
--(unsigned)hash;
+-(unsigned long)hash;
 
 @end
